+++ crates/full-node/sov-sequencer/src/preferred/cache_warm_up_executor.rs
@@ -60,7 +60,7 @@ impl<S: Spec> Clone for StartBlockNotification<S> {
             data: self.data.clone(),
             checkpoint: self
                 .checkpoint
-                .clone_with_empty_witness_dropping_temp_cache(),
+                .clone_with_empty_witness_dropping_temp_cache_and_ignoring_pinned_cache(),
             sequence_number: self.sequence_number,
         }
     }
@@ -209,6 +209,7 @@ impl<S: Spec> CacheWarmUpExecutor<S> {
                 exec_config,
                 seq_config.clone(),
                 Default::default(),
+                None, // TODO: Consider adding a pinned cache to the warmup executors
             );
 
             let mut maybe_executor_sequence_number = None;