+++ crates/full-node/sov-sequencer/src/preferred/state_root_compute.rs
@@ -186,7 +186,7 @@ async fn compute_state_root<S: Spec, Rt: Runtime<S>>(
                 }
 
                 storage
-                    .compute_state_update(state_accesses, &Default::default(), prev_root)
+                    .compute_state_update(state_accesses, &Default::default(), prev_root, None)
                     .expect("Failed to compute state update").0
             })
     });
@@ -411,7 +411,7 @@ mod tests {
     ) -> Arc<RawStateChanges> {
         let mut rt = Rt::default();
         let mut kernel = rt.kernel();
-        let mut checkpoint = StateCheckpoint::new(storage.clone(), &kernel);
+        let mut checkpoint = StateCheckpoint::new(storage.clone(), &kernel, None);
         let mut state_with_partially_stale_heights =
             KernelStateAccessor::from_checkpoint(&kernel, &mut checkpoint);
         let height = header.height;
@@ -455,7 +455,7 @@ mod tests {
     ) -> Arc<RawStateChanges> {
         let mut rt = Rt::default();
         let mut kernel = rt.kernel();
-        let mut checkpoint = StateCheckpoint::new(storage.clone(), &kernel);
+        let mut checkpoint = StateCheckpoint::new(storage.clone(), &kernel, None);
         let mut state_with_partially_stale_heights =
             KernelStateAccessor::from_checkpoint(&kernel, &mut checkpoint);
         let height = header.height;
@@ -571,7 +571,7 @@ mod tests {
         let writes_on_the_node = writes_only_kernel::<S, Rt>(&node_storage);
         let prev_root = <S::Storage as Storage>::PRE_GENESIS_ROOT;
         let (node_new_root, changes) = node_storage
-            .compute_state_update(writes_on_the_node, &Default::default(), prev_root)
+            .compute_state_update(writes_on_the_node, &Default::default(), prev_root, None)
             .unwrap();
         storage_manager.commit_state_update(node_storage, changes, node_new_root);
     }
@@ -600,6 +600,7 @@ mod tests {
                 writes_on_the_node.to_state_accesses_for_sequencer_state_root_computation(),
                 &Default::default(),
                 prev_root,
+                None,
             )
             .unwrap();
 
@@ -678,7 +679,7 @@ mod tests {
         let writes_on_the_node = writes_only_kernel::<S, Rt>(&node_storage);
         let mut prev_root = <S::Storage as Storage>::PRE_GENESIS_ROOT;
         let (node_new_root, changes) = node_storage
-            .compute_state_update(writes_on_the_node, &Default::default(), prev_root)
+            .compute_state_update(writes_on_the_node, &Default::default(), prev_root, None)
             .unwrap();
         prev_root = node_new_root;
         let to_commit = node_storage.materialize_changes(changes);
@@ -716,6 +717,7 @@ mod tests {
                     raw_state_changes.to_state_accesses_for_sequencer_state_root_computation(),
                     &Default::default(),
                     prev_root,
+                    None,
                 )
                 .unwrap();
             changes.add_accessory_items(
@@ -836,6 +838,7 @@ mod tests {
                 writes_on_the_node.to_state_accesses_for_sequencer_state_root_computation(),
                 &Default::default(),
                 prev_root,
+                None,
             )
             .unwrap();