+++ crates/module-system/sov-modules-api/src/state/accessors/checkpoints.rs
@@ -1,5 +1,6 @@
 use sov_metrics::{StateAccessMetric, StateMetrics};
 use sov_rollup_interface::common::{SlotNumber, VisibleSlotNumber};
+use sov_state::pinned_cache::PinnedCache;
 #[cfg(feature = "native")]
 use sov_state::StateGetter;
 use sov_state::{IsValueCached, Namespace, SlotKey, SlotValue, StateAccesses, Storage};
@@ -92,7 +93,7 @@ impl<S: Spec> StateCheckpoint<S> {
     /// such as in the API accessors.
     #[must_use]
     #[cfg(feature = "native")]
-    pub fn clone_with_empty_witness_dropping_temp_cache(&self) -> Self {
+    pub fn clone_with_empty_witness_dropping_temp_cache_and_ignoring_pinned_cache(&self) -> Self {
         Self {
             delta: self.delta.clone_with_empty_witness(),
             visible_slot_num: self.visible_slot_num,
@@ -102,19 +103,55 @@ impl<S: Spec> StateCheckpoint<S> {
         }
     }
 
+    /// Deep copy the state checkpoint (including its state caches), ignoring
+    /// the witness and the temp cache.
+    ///
+    /// Since this method leaves the witness of the new
+    /// checkpoint in a state that is inconsistent with its caches,
+    /// it should only be used in situations where the witness is not needed,
+    /// such as in the API accessors.
+    #[must_use]
+    #[cfg(feature = "native")]
+    pub fn clone_with_empty_witness_dropping_temp_cache_but_taking_pinned_cache(&mut self) -> Self {
+        let pinned_cache = self.take_pinned_cache();
+        let mut new_checkpoint = Self {
+            delta: self.delta.clone_with_empty_witness(),
+            visible_slot_num: self.visible_slot_num,
+            rollup_height: self.rollup_height,
+            cache: TempCache::new(),
+            metrics: StateMetrics::default(),
+        };
+        new_checkpoint.set_pinned_cache(pinned_cache);
+        new_checkpoint
+    }
+
     /// Creates a new [`StateCheckpoint`] instance with the given intermediate state that will be checked before storage when a value isn't already present in the checkpoint.
     #[cfg(feature = "native")]
     pub fn new_with_uncomitted_changes<K: Kernel<S>>(
         inner: S::Storage,
         kernel: &K,
         uncomitted_changes: Box<dyn StateGetter>,
+        pinned_cache: Option<PinnedCache>,
     ) -> Self {
         Self::with_witness_and_uncomitted_changes(
             inner,
             Default::default(),
             kernel,
             Some(uncomitted_changes),
+            pinned_cache,
         )
+    }
+
+    #[cfg(feature = "native")]
+    /// Takes the pinned user-state cache. See  [`PinnedCache`] for more details.
+    pub fn take_pinned_cache(&mut self) -> Option<PinnedCache> {
+        self.delta.user_cache.take_pinned_cache()
+    }
+
+    #[cfg(feature = "native")]
+    /// Sets the pinned cache for the state checkpoint. See [`PinnedCache`] for more details.
+    pub fn set_pinned_cache(&mut self, pinned_cache: Option<PinnedCache>) {
+        self.delta.user_cache.set_pinned_cache(pinned_cache);
     }
 
     /// Replace the storage and intermediate state underlying the checkpoint in place. It is up to the caller
@@ -137,14 +174,18 @@ impl<S: Spec> StateCheckpoint<S> {
         &self,
         kernel: &K,
     ) -> RollupHeight {
-        let new_checkpoint = Self::new(self.delta.inner().clone(), kernel);
+        let new_checkpoint = Self::new(self.delta.inner().clone(), kernel, None);
         new_checkpoint.rollup_height
     }
 
     /// Creates a new [`StateCheckpoint`] instance without any changes, backed
     /// by the given [`Storage`].
-    pub fn new<K: Kernel<S>>(inner: S::Storage, kernel: &K) -> Self {
-        Self::with_witness(inner, Default::default(), kernel)
+    pub fn new<K: Kernel<S>>(
+        inner: S::Storage,
+        kernel: &K,
+        pinned_cache: Option<PinnedCache>,
+    ) -> Self {
+        Self::with_witness(inner, Default::default(), kernel, pinned_cache)
     }
 
     /// Creates a new [`StateCheckpoint`] instance without any changes, backed
@@ -153,6 +194,7 @@ impl<S: Spec> StateCheckpoint<S> {
         inner: S::Storage,
         witness: <S::Storage as Storage>::Witness,
         kernel: &K,
+        pinned_cache: Option<PinnedCache>,
     ) -> Self {
         Self::with_witness_and_uncomitted_changes(
             inner,
@@ -160,6 +202,7 @@ impl<S: Spec> StateCheckpoint<S> {
             kernel,
             #[cfg(feature = "native")]
             None,
+            pinned_cache,
         )
     }
 
@@ -170,11 +213,13 @@ impl<S: Spec> StateCheckpoint<S> {
         witness: <S::Storage as Storage>::Witness,
         kernel: &K,
         #[cfg(feature = "native")] uncomitted_changes: Option<Box<dyn StateGetter>>,
+        #[cfg_attr(not(feature = "native"), allow(unused))] pinned_cache: Option<PinnedCache>,
     ) -> Self {
         let mut delta = Delta::with_witness(inner, witness);
         #[cfg(feature = "native")]
         {
             delta.uncomitted_changes = uncomitted_changes;
+            delta.user_cache.set_pinned_cache(pinned_cache);
         }
         let mut metrics = StateMetrics::default();
         let mut bootstrap_state = BootstrapWorkingSet {
@@ -224,7 +269,7 @@ impl<S: Spec> StateCheckpoint<S> {
     /// them to compute the `StateUpdate` created by this `StateCheckpoint`.
     #[allow(clippy::type_complexity)]
     pub fn materialize_update(
-        self,
+        #[cfg_attr(not(feature = "native"), allow(unused_mut))] mut self,
         prev_state_root: <S::Storage as Storage>::Root,
     ) -> (
         <S::Storage as Storage>::Root,
@@ -233,10 +278,20 @@ impl<S: Spec> StateCheckpoint<S> {
         <S::Storage as Storage>::Witness,
         S::Storage,
     ) {
+        let pinned_cache = {
+            #[cfg(feature = "native")]
+            {
+                self.take_pinned_cache()
+            }
+            #[cfg(not(feature = "native"))]
+            {
+                None
+            }
+        };
         let (cache_log, accessory_delta, witness, storage) = self.delta.freeze();
         let _span = tracing::debug_span!("compute_state_root", scope = "node").entered();
         let (root, update) = storage
-            .compute_state_update(cache_log, &witness, prev_state_root)
+            .compute_state_update(cache_log, &witness, prev_state_root, pinned_cache)
             .expect("state update computation must succeed");
         tracing::trace!(%root, "computed state root");
         (root, update, accessory_delta, witness, storage)