+++ crates/module-system/sov-modules-api/tests/integration/state_tests/archival.rs
@@ -33,7 +33,7 @@ where
 
     for current_height in 0..100 {
         let (storage, prev_root) = storage_manager.create_storage_with_root();
-        let state_checkpoint = StateCheckpoint::new(storage.clone(), &kernel);
+        let state_checkpoint = StateCheckpoint::new(storage.clone(), &kernel, None);
         let api_accessor = ApiStateAccessor::new(&state_checkpoint, Arc::new(kernel.clone()));
 
         for past_height in 0..current_height {
@@ -54,7 +54,7 @@ where
             prev_root,
         );
         let storage = storage_manager.create_prover_storage();
-        let state_checkpoint = StateCheckpoint::new(storage, &kernel);
+        let state_checkpoint = StateCheckpoint::new(storage, &kernel, None);
         let api_accessor = ApiStateAccessor::new(&state_checkpoint, Arc::new(kernel.clone()));
 
         for another_past_height in 0..=current_height {