+++ crates/module-system/sov-modules-macros/tests/integration/dispatch.rs
@@ -315,8 +315,11 @@ mod derive_genesis {
     #[test]
     fn derive_genesis() {
         let storage = ZkStorage::new();
-        let mut state =
-            sov_modules_api::StateCheckpoint::new(storage, &MockKernel::<ZkTestSpec>::default());
+        let mut state = sov_modules_api::StateCheckpoint::new(
+            storage,
+            &MockKernel::<ZkTestSpec>::default(),
+            None,
+        );
         let runtime = &mut Runtime::<ZkTestSpec, u32>::default();
         let config = GenesisConfig::new(Default::default(), (), ());
         let mut genesis_state =
@@ -378,8 +381,11 @@ mod derive_dispatch {
 
         let storage = ZkStorage::new();
 
-        let mut state =
-            sov_modules_api::StateCheckpoint::new(storage, &MockKernel::<ZkTestSpec>::default());
+        let mut state = sov_modules_api::StateCheckpoint::new(
+            storage,
+            &MockKernel::<ZkTestSpec>::default(),
+            None,
+        );
         let config = GenesisConfig::new(Default::default(), (), ());
         let mut genesis_state =
             state.to_genesis_state_accessor::<Runtime<ZkTestSpec, u32>>(&config);