+++ crates/module-system/sov-modules-rollup-blueprint/src/native_only/mod.rs
@@ -442,8 +442,7 @@ pub trait FullNodeBlueprint<M: ExecutionMode>: RollupBlueprint<M> {
                 .await?;
 
         let mut rt = Self::Runtime::default();
-
-        let checkpoint = StateCheckpoint::new(prover_storage, &rt.kernel());
+        let checkpoint = StateCheckpoint::new(prover_storage, &rt.kernel(), None);
         let current_height = checkpoint.rollup_height_to_access();
 
         validate_heights(