+++ crates/module-system/sov-state/src/lib.rs
@@ -9,6 +9,7 @@ pub mod config;
 mod event;
 pub mod namespaces;
 pub mod nomt;
+pub mod pinned_cache;
 #[cfg(feature = "native")]
 mod prover_storage;
 #[cfg(feature = "native")]