+++ crates/module-system/sov-test-utils/src/runtime/mod.rs
@@ -341,7 +341,7 @@ where
         let mut runtime = RT::default();
         let kernel = runtime.kernel();
 
-        let mut state_checkpoint = StateCheckpoint::<S>::new(stf_state.clone(), &kernel);
+        let mut state_checkpoint = StateCheckpoint::<S>::new(stf_state.clone(), &kernel, None);
 
         let base_fee_per_gas = RT::default()
             .chain_state()
@@ -362,7 +362,7 @@ where
         let mut runtime = RT::default();
         let kernel = runtime.kernel();
 
-        let mut state_checkpoint = StateCheckpoint::<S>::new(stf_state.clone(), &kernel);
+        let mut state_checkpoint = StateCheckpoint::<S>::new(stf_state.clone(), &kernel, None);
 
         let base_fee_per_gas = RT::default()
             .chain_state()
@@ -425,7 +425,7 @@ where
 
         let mut runtime = RT::default();
 
-        let mut state = StateCheckpoint::<S>::new(stf_state.clone(), &runtime.kernel());
+        let mut state = StateCheckpoint::<S>::new(stf_state.clone(), &runtime.kernel(), None);
 
         let mut kernel_state = runtime.kernel().accessor(&mut state);
 
@@ -456,7 +456,7 @@ where
     fn synchronize_storage_channel(&mut self) {
         let storage = self.storage_manager.create_prover_storage();
         self.checkpoint_sender
-            .send(Arc::new(StateCheckpoint::new(storage, &RT::default().kernel())))
+            .send(Arc::new(StateCheckpoint::new(storage, &RT::default().kernel(), None)))
             .expect("Failed to send storage, the storage channel is closed. This is a bug. Please report it.");
     }
 
@@ -481,6 +481,7 @@ where
         let (sender, receiver) = watch::channel(Arc::new(StateCheckpoint::new(
             stf_state.clone(),
             &RT::default().kernel(),
+            None,
         )));
 
         let (state_root, change_set) =