+++ crates/module-system/sov-test-utils/src/test_rollup.rs
@@ -14,7 +14,7 @@ use crate::postgres::PostgresImage;
 use crate::{Transaction, TEST_MOCK_DA_POLLING_INTERVAL};
 use crate::{
     TEST_DEFAULT_PROVER_ADDRESS, TEST_DEFAULT_SEQUENCER_ADDRESS, TEST_MAX_BATCH_SIZE,
-    TEST_MAX_CONCURRENT_BLOBS, TEST_NUM_CACHE_WARMUP_WORKERS,
+    TEST_MAX_CONCURRENT_BLOBS,
 };
 use anyhow::Context;
 use derivative::Derivative;
@@ -113,7 +113,6 @@ pub struct RollupBuilderConfig<S: Spec> {
     pub start_at_rollup_height: Option<RollupHeight>,
     pub stop_at_rollup_height: Option<RollupHeight>,
     pub extension: Option<SeqConfigExtension>,
-    pub num_cache_warmup_workers: usize,
     pub separate_archival_db: bool,
 }
 
@@ -332,6 +331,7 @@ impl<R: FullNodeBlueprint<Native> + Default + 'static> RollupBuilder<R> {
         if self.config.separate_archival_db {
             rollup_db_config.separate_archival_state = true;
         }
+
         RollupConfig {
             storage: rollup_db_config,
             runner: RunnerConfig {
@@ -412,7 +412,6 @@ impl<R: FullNodeBlueprint<Native> + Default + 'static> RollupBuilder<R> {
                 max_log_limit: 20000,
                 response_size_limit: (1024 * 1024) - (1024 * 30), // Limit our response size to 1MB, leaving 30kb for headers, overhead, and misestimation.
             }),
-            num_cache_warmup_workers: TEST_NUM_CACHE_WARMUP_WORKERS,
             separate_archival_db: true,
         }
     }